% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{calc_power_common_rare}
\alias{calc_power_common_rare}
\title{Power calculation for a given study design}
\usage{
calc_power_common_rare(N_sample = 200, N_covariate = 8, N_feature = 300,
  N_repeat = 1, conti_prop = 0.5, pos_prop = 0.2, contin = "continuous",
  increment = 0.005, q_cutoff = 0.05, quartile_common, quartile_rare,
  v_mean_transform)
}
\arguments{
\item{N_sample}{Number of subjects in the study. Integer.}

\item{N_covariate}{Number of covariates in the study. Integer.}

\item{N_feature}{Number of expected species or pathways. Integer.}

\item{N_repeat}{Number of repeat measurements for each data set. Integer.}

\item{conti_prop}{The proportion of continuous covariates. A number in [0,1]. Default is 0.5.}

\item{pos_prop}{The proportion of species or pathways that are truely correlated with the covariate we are interested in. A number in [0,1]. Default is 0.2.}

\item{contin}{The type of covariate we are interested. The type includeds "continous", "discrete" and "inter" (interaction). Strint. Default is "continuous".}

\item{increment}{A small number to adjust the smoothness of the power curve. The smaller, the more smooth. Default is 0.005.}

\item{q_cutoff}{The threshold for q value (p value after BH adjustment). A number in (0,1). Default is 0.05.}

\item{quartile_common}{Standard deviation quartile common}

\item{quartile_rare}{Standard deviation quartile rare}

\item{v_mean_transform}{Transformed mean distribution}
}
\value{
A data frame with power and beta.
}
\description{
\code{calc_power_common_rare} calculates the power for common and rare species and pathways. The difference between this function and \code{calc_power} is that \code{calc_power} does not consider common and rare features.
}
\details{
Returning a data frame containing power, FPR, Beta and mininum difference
}
\author{
Liu Cao
}
