% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_ref_distribution}
\alias{calc_ref_distribution}
\title{Generate mean distribution and quartiles of standard deviation}
\usage{
calc_ref_distribution(file, common_rare = TRUE, sep = "\\t")
}
\arguments{
\item{file}{The full path of the reference data. String.}

\item{common_rare}{An indicator whether to distinguish the features to be common and rare features. Bool. Default is TRUE.}

\item{sep}{The seperator of the reference data. String. Default is \\t.}
}
\value{
A list of two or three element: mean_distribution and (sd_quartile or sd_quartile_common, sd_quartile_rare).
}
\description{
\code{calc_ref_distribution} calculate reference distribution
}
\details{
Given a reference dataset, perform arcsin square root transformation, and calculate its mean distribution, standard deviation of common and rare.
The reference dataset should contain relative abundance.
}
\author{
Liu Cao
}
