% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_power}
\alias{plot_power}
\title{Plot power and fpr curve}
\usage{
plot_power(power.obj, target_power = 0.9, target_fpr = 0.05, title = "")
}
\arguments{
\item{power.obj}{A data frame generated by \code{calc_power} or \code{calc_power_common_rare}}

\item{target_power}{The least power the study wants to reach. A number in (0,1). Default is 0.9.}

\item{target_fpr}{The highest fpr of the study. A number in (0,1). Default is 0.05.}

\item{title}{title for the plot. Default is "".}
}
\value{
None.
}
\description{
\code{plot_power} plots the power and fpr curve for common and rare species or pathways.
}
\details{
ehehehehheheheheheheheeh
}
\author{
Liu Cao
}
