% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{result_power}
\alias{result_power}
\title{Minimal difference for target power}
\usage{
result_power(power.obj, target_power = 0.9, target_fpr = 0.05)
}
\arguments{
\item{power.obj}{A data frame generated by \code{calc_power}}

\item{target_power}{The least power the study wants to reach. A number in (0,1). Default is 0.9.}

\item{target_fpr}{The highest fpr of the study. A number in (0,1). Default is 0.05.}

\item{title}{title for the plot. Default is "".}
}
\value{
None.
}
\description{
\code{result_power} Calculate minimal difference for common and rare species or pathways at target power.
}
\details{
ehehehehheheheheheheheeh
}
\author{
Liu Cao
}
